suppression de validDelai.ts car validation prise en charge par PR #143
